Correct dependency injection
